Created test folder
moved all tests into a folder and created a test suite to run all tests
