+++ Solitare/CardDeck.java
@@ -1,3 +1,4 @@
+package Solitare;
 /**
  * card deck class
  *
