+++ Solitare/driver.java
@@ -1,3 +1,4 @@
+package Solitare;
 import java.util.ArrayList;
 import java.util.Stack;
